Code Refactoring
Refactored the code by moving the logic from App.js to UsersChart.js
